scripts/requirements.txt
numpy # add version number
opencv-python
matplotlib
openpyxl
pillow
pandas
pyqt5
